API Documentation

Thermodynamics

Breeze.Thermodynamics.AtmosphereThermodynamics — Type

AtmosphereThermodynamics(FT = Oceananigans.defaults.FloatType;
                         gravitational_acceleration = 9.81,
                         molar_gas_constant = 8.314462618,
                         dry_air_molar_mass = 0.02897,
                         dry_air_heat_capacity = 1005,
                         vapor_molar_mass = 0.018015,
                         vapor_heat_capacity = 1850,
                         saturation = Saturation(FT),
                         condensation = water_condensation(FT),
                         deposition = water_deposition(FT))

Create AtmosphereThermodynamics with parameters that correpsond to the composition of dry air in Earth's atmosphere and water vapor. The default isnothing(condensation) implies thermodynamics appropriate for unsaturated and therefore non-condensing air.

Breeze.Thermodynamics.saturation_vapor_pressure — Function

saturation_vapor_pressure(T, thermo)

Compute the saturation vapor pressure over a liquid surface by integrating the Clausius-Clapeyron relation,

\[dp/dT = ℒᵛ / (Rᵛ T^2)\]